fixed datatypes
- removed all float values (only integer is needed)
- changed standard values (-1,-1.0,"undefined",False)

+++ src/database/eRepDB.py
@@ -42,7 +42,7 @@ class eRepLokalDBManager():
                 citPartyName TEXT,
                 citPartyID INTEGER,
                 citMilitarySkillLevel TEXT,
-                citMilitarySkillPoints REAL,
+                citMilitarySkillPoints INTEGER,
                 citMilitaryRankPoints INTEGER,
                 citMilitaryRank TEXT,
                 citMilitaryStars INTEGER,
